refine allocated memory checking to accommodate more pages (#184)
* refine allocated memory checking to accommodate more pages

* fix: make sure effective block index is less than allocated pages

* chore: fix code style

* chore: fix code style

+++ crates/specs/src/mtable.rs
@@ -90,7 +90,7 @@ impl From<crate::types::ValueType> for VarType {
 }
 
 impl MemoryReadSize {
-    pub fn byte_size(&self) -> u64 {
+    pub fn byte_size(&self) -> u32 {
         match self {
             MemoryReadSize::U8 => 1,
             MemoryReadSize::S8 => 1,
